requirements.txt
numpy==1.19.2
torchvision==0.8.2
click==7.1.2
scipy==1.6.0
torch==1.7.1
Pillow==8.2.0
pycuda==2021.1
tensorrt==0.0.1
tools==0.1.9
